# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/membership": {
            "get": {
                "description": "Get a list of Membership.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Membership"
                ],
                "summary": "Get All Membership.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Membership"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Membership.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Membership"
                ],
                "summary": "Create New Membership.",
                "parameters": [
                    {
                        "description": "The body to create a new Membership",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.membershipInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Membership"
                        }
                    }
                }
            }
        },
        "/membership/{id}": {
            "get": {
                "description": "Get a Membership by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Membership"
                ],
                "summary": "Get Membership by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Membership"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Membership by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Membership"
                ],
                "summary": "Delete one Membership.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Membership ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Membership by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Membership"
                ],
                "summary": "Update Membership.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Membership ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The body to update an Membership",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.membershipInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Membership"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controllers.membershipInput": {
            "type": "object",
            "required": [
                "discount",
                "membership_name"
            ],
            "properties": {
                "discount": {
                    "type": "integer"
                },
                "membership_name": {
                    "type": "string"
                }
            }
        },
...
        "models.Membership": {
            "type": "object",
            "required": [
                "discount",
                "membership_name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "discount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "membership_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}